requirements.txt
pandas==2.2.3
protobuf==3.20.3
python-dotenv==1.1.0
Requests==2.32.3
